backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-Mail==0.9.1
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==21.2.0
email-validator==2.1.0
Werkzeug==3.0.1
Jinja2==3.1.3
SQLAlchemy==2.0.23
alembic==1.13.1
